// src/app/public/models/ROM/entel/validacion-bundles/validacionbundle.ts
export class ValidacionBundle{
    intidventasprincipal?: number;
    intventasromid?: number;
    strdtevestasromfeope?: string;
    strdnicliente?: string;
    strventasromusucr?: string;
    nombrepromotor?: string;
    intproductoid?: number;
    strproductodesc?: string;
    intplanid?: number;
    strplandesc?: string;
    intmodeloequipoid?: number;
    strmodeloequipodesc?: string;
    intbundleid?: number;
    descripcion?: string;
}